pyproject.toml
[project]
name = "sundai"
version = "0.1.0"
description = "Automated social media posting using Notion, OpenRouter LLM, and Mastodon"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
